+++ custom_components/solarcharger/models/model_allocation.py
@@ -17,6 +17,8 @@ class PowerAllocation:
     max_power: float
     # Not used. FYI only.
     max_current: float
+    # Charger step power list.
+    step_power_list: list[float]
     # Minimum power required for the charger to operate. Must be -ve for surplus power.
     activation_power: float
     adjusted_activation_power: float
@@ -64,6 +66,8 @@ class PowerAllocation:
     # final_power is used by paused chargers to determine when to exit paused state.
     allocation_final_weight: float = 0.0  # Use this weight for allocation.
     deallocation_final_weight: float = 0.0  # Use this weight for deallocation.
+
+    # Net allocated power after rebalancing. -ve = allocation. +ve = deallocation.
     final_power: float = 0.0
 
     # ----------------------------------------------------------------------------
@@ -73,6 +77,7 @@ class PowerAllocation:
             f"name={self.name}, "
             f"max_power={self.max_power}, "
             f"max_current={self.max_current}, "
+            f"step_power_list={self.step_power_list}, "
             f"adjusted_activation_power={self.adjusted_activation_power} ("
             f"{self.activation_power}), "
             f"priority={self.priority}, "